git-debimport zookeeper_3.4.9-3+deb8u1.debian.tar.xz

+++ src/java/main/org/apache/zookeeper/server/persistence/FileTxnLog.java
@@ -111,7 +111,11 @@ public class FileTxnLog implements TxnLog {
                 LOG.warn(size + " is not a valid value for preAllocSize");
             }
         }
-        fsyncWarningThresholdMS = Long.getLong("fsync.warningthresholdms", 1000);
+        /** Local variable to read fsync.warningthresholdms into */
+        Long fsyncWarningThreshold;
+        if ((fsyncWarningThreshold = Long.getLong("zookeeper.fsync.warningthresholdms")) == null)
+            fsyncWarningThreshold = Long.getLong("fsync.warningthresholdms", 1000);
+        fsyncWarningThresholdMS = fsyncWarningThreshold;
     }
 
     long lastZxidSeen;
@@ -285,9 +289,10 @@ public class FileTxnLog implements TxnLog {
         // if a log file is more recent we must scan it to find
         // the highest zxid
         long zxid = maxLog;
+        TxnIterator itr = null;
         try {
             FileTxnLog txn = new FileTxnLog(logDir);
-            TxnIterator itr = txn.read(maxLog);
+            itr = txn.read(maxLog);
             while (true) {
                 if(!itr.next())
                     break;
@@ -296,8 +301,20 @@ public class FileTxnLog implements TxnLog {
             }
         } catch (IOException e) {
             LOG.warn("Unexpected exception", e);
+        } finally {
+            close(itr);
         }
         return zxid;
+    }
+
+    private void close(TxnIterator itr) {
+        if (itr != null) {
+            try {
+                itr.close();
+            } catch (IOException ioe) {
+                LOG.warn("Error closing file iterator", ioe);
+            }
+        }
     }
 
     /**
@@ -347,17 +364,27 @@ public class FileTxnLog implements TxnLog {
      * @return true if successful false if not
      */
     public boolean truncate(long zxid) throws IOException {
-        FileTxnIterator itr = new FileTxnIterator(this.logDir, zxid);
-        PositionInputStream input = itr.inputStream;
-        long pos = input.getPosition();
-        // now, truncate at the current position
-        RandomAccessFile raf=new RandomAccessFile(itr.logFile,"rw");
-        raf.setLength(pos);
-        raf.close();
-        while(itr.goToNextLog()) {
-            if (!itr.logFile.delete()) {
-                LOG.warn("Unable to truncate " + itr.logFile);
+        FileTxnIterator itr = null;
+        try {
+            itr = new FileTxnIterator(this.logDir, zxid);
+            PositionInputStream input = itr.inputStream;
+            if(input == null) {
+                throw new IOException("No log files found to truncate! This could " +
+                        "happen if you still have snapshots from an old setup or " +
+                        "log files were deleted accidentally or dataLogDir was changed in zoo.cfg.");
             }
+            long pos = input.getPosition();
+            // now, truncate at the current position
+            RandomAccessFile raf = new RandomAccessFile(itr.logFile, "rw");
+            raf.setLength(pos);
+            raf.close();
+            while (itr.goToNextLog()) {
+                if (!itr.logFile.delete()) {
+                    LOG.warn("Unable to truncate {}", itr.logFile);
+                }
+            }
+        } finally {
+            close(itr);
         }
         return true;
     }
@@ -627,6 +654,9 @@ public class FileTxnLog implements TxnLog {
                 }
                 // if we went to the next log file, we should call next() again
                 return next();
+            } catch (IOException e) {
+                inputStream.close();
+                throw e;
             }
             return true;
         }
@@ -654,7 +684,9 @@ public class FileTxnLog implements TxnLog {
          * and release the resources.
          */
         public void close() throws IOException {
-            inputStream.close();
+            if (inputStream != null) {
+                inputStream.close();
+            }
         }
     }
